refactor(frontend): extract run-page constants and helpers to run-config/run-helpers
